Refactor/회원가입 dropdown 컴포넌트 파일로 생성
